+++ app/components/rad-card/body.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-radical/components/rad-card/body';